requirements.txt
python-dotenv >=0.18.0
PySimpleGUI >= 4.45.0